requirements.txt
customtkinter
pillow
psutil
pyautogui
pyperclip
pywin32
cryptography